wip Try use TxBody pattern synonym to no success
Wrote this on github as comment:

It seems that the scripts I'm picking out and using are not included in the pattern TxBody: https://github.com/input-output-hk/cardano-node/blob/master/cardano-api/src/Cardano/Api/TxBody.hs#L2070

Question is, if the scripts that I pick out are the right ones? As I found also that inside TxBody the Script type also exist under:

    TxBodyContent{txOuts} -> TxOut -> ReferenceScript -> ScriptInAnyLang
    TxBodyContent{txReturnCollateral} -> TxReturnCollateral -> TxOut
    TxBodyContent{txAuxScripts} -> TxAuxScripts -> ScriptInEra

I would guess the scripts at the inputs (the ones that I'm using already and that come in as serialised) are the important ones, as those are the ones that are run.